Quota: use a big integer
In Django IntegerField are limited to [-2147483648, 2147483647] which might not
be enough (around 2GB).
Using BigIntegerField allow for a quota up to 9223372036854775807.

Thanks to Dean Birch for the bug report.

+++ Artifactorial/models.py
@@ -53,8 +53,8 @@ class Directory(models.Model):
     is_public = models.BooleanField(default=False)
     ttl = models.IntegerField(blank=False, default=90,
                               help_text="Files TTL in days")
-    quota = models.IntegerField(blank=False, default=1024*1024*1024,
-                                help_text='Size limit in Bytes')
+    quota = models.BigIntegerField(blank=False, default=1024*1024*1024,
+                                   help_text='Size limit in Bytes')
 
     class Meta:
         verbose_name_plural = 'Directories'
